# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MiniShop API",
        "description": "API documentation for the MiniShop e-commerce backend.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Current server"
        }
    ],
    "paths": {
...
        "/api/banners": {
            "get": {
                "tags": [
                    "Banners"
                ],
                "summary": "List all banners",
                "operationId": "9f718c7db0d13e4b5be26b80595f4cf1",
                "responses": {
                    "200": {
                        "description": "Banners retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/SuccessResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Banner"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Banners"
                ],
                "summary": "Create a new banner (Admin only)",
                "operationId": "546707dc3442c51e694ec312df3664e9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "image"
                                ],
                                "properties": {
                                    "image": {
                                        "description": "Image file (max 5MB)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "alt": {
                                        "type": "string",
                                        "example": "Summer sale banner",
                                        "nullable": true
                                    },
                                    "url": {
                                        "type": "string",
                                        "example": "/products/wireless-mouse",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Banner created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/SuccessResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/Banner"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Not an Admin"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/banners/{id}": {
            "get": {
                "tags": [
                    "Banners"
                ],
                "summary": "Get a single banner",
                "operationId": "13d01a36bda0d8b491a382619c6bc469",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Banner retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/SuccessResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/Banner"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Banner not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Banners"
                ],
                "summary": "Update a banner (Admin only)",
                "operationId": "89bb1eb70cfa7909040b083ebe79f6df",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "_method": {
                                        "description": "Laravel method spoofing (required since file uploads need POST)",
                                        "type": "string",
                                        "example": "PUT"
                                    },
                                    "image": {
                                        "description": "Leave empty to keep the current image",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "alt": {
                                        "type": "string",
                                        "example": "Summer sale banner",
                                        "nullable": true
                                    },
                                    "url": {
                                        "type": "string",
                                        "example": "/products/wireless-mouse",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Banner updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/SuccessResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/Banner"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Not an Admin"
                    },
                    "404": {
                        "description": "Banner not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Banners"
                ],
                "summary": "Delete a banner (Admin only)",
                "operationId": "00a27f787ec0124f0b98cfdeecf6090a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Banner deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Not an Admin"
                    },
                    "404": {
                        "description": "Banner not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "SuccessResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Success"
                    },
                    "data": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "Banner": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "image": {
                        "type": "string",
                        "example": "http://localhost:8000/storage/banners/abc.jpg"
                    },
                    "alt": {
                        "type": "string",
                        "example": "Summer sale banner",
                        "nullable": true
                    },
                    "url": {
                        "type": "string",
                        "example": "/products/wireless-mouse",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter the token obtained from the login endpoint, without the \"Bearer\" prefix.",
                "bearerFormat": "Sanctum token",
                "scheme": "bearer"
            }
        }
    },
...
}